Fixes requests by mantainer.
