+++ Platform.App/python-template/core/temporal/utils.py
@@ -60,3 +60,23 @@ def truncate_identifier(identifier):
                           sa_util.md5_hex(identifier)[-4:])
 
     return identifier
+
+def addTemporal(session,instances):
+    temporal_structures_clock = []
+    temporal_structures_history = []
+
+    for entity in instances:
+        clock, _ = session.create_clock_bulk(entity)
+
+        for field in entity.Temporal.fields:
+            history, history_created = session.create_field_history_bulk(
+            entity, field, clock,None)
+            temporal_structures_history.append(history)
+
+        clock.ticks = 1
+        temporal_structures_clock.append(clock)
+
+    session.bulk_save_objects(temporal_structures_clock)
+    session.flush()
+    session.bulk_save_objects(temporal_structures_history)
+    session.flush()
